+++ src/something_main.cpp
@@ -156,7 +156,7 @@ int main(void)
     float lag_sec = 0;
     float next_sec = 0;
     Uint32 frames_of_current_second = 0;
-    Uint32 fps = 0;
+    Uint16 fps = 0;
     while (!game.quit) {
         Uint32 curr_ticks = SDL_GetTicks();
         float elapsed_sec = (float) (curr_ticks - prev_ticks) / 1000.0f;
